+++ app/src/main/java/porce/com/porce/fragments/MyPageFragment.java
@@ -1,0 +1,38 @@
+package porce.com.porce.fragments;
+
+
+import android.os.Bundle;
+import android.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import porce.com.porce.R;
+
+/**
+ * A simple {@link Fragment} subclass.
+ * Use the {@link MyPageFragment#newInstance} factory method to
+ * create an instance of this fragment.
+ */
+public class MyPageFragment extends Fragment {
+
+    /**
+     * Use this factory method to create a new instance of
+     * this fragment using the provided parameters.
+     *
+     * @return A new instance of fragment MyPageFragment.
+     */
+    public static MyPageFragment newInstance() {
+        MyPageFragment fragment = new MyPageFragment();
+        return fragment;
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_my_page, container, false);
+    }
+
+
+}
