+++ app/src/main/res/drawable/fab_label_background.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <solid android:color="@color/black_semi_transparent"/>
+    <padding
+        android:left="16dp"
+        android:top="4dp"
+        android:right="16dp"
+        android:bottom="4dp"/>
+    <corners
+        android:radius="2dp"/>
+</shape>
