+++ app/src/main/res/values/strings.xml
@@ -5,4 +5,9 @@
     <string name="hello_world">Hello world!</string>
     <string name="action_settings">Settings</string>
 
+    <!-- floating action button -->
+    <string name="floating_button_label_post">本紹介</string>
+    <string name="floating_button_label_find">本探し</string>
+    <string name="hello_blank_fragment">Hello blank fragment</string>
+
 </resources>
